// dss-web/src/app/components/dataset/models/richDatasetModel.ts
/*
 *   HORTONWORKS DATAPLANE SERVICE AND ITS CONSTITUENT SERVICES
 *
 *   (c) 2016-2018 Hortonworks, Inc. All rights reserved.
 *
 *   This code is provided to you pursuant to your written agreement with Hortonworks, which may be the terms of the
 *   Affero General Public License version 3 (AGPLv3), or pursuant to a written agreement with a third party authorized
 *   to distribute this code.  If you do not have a written agreement with Hortonworks or with an authorized and
 *   properly licensed third party, you do not have any rights to this code.
 *
 *   If this code is provided to you under the terms of the AGPLv3:
 *   (A) HORTONWORKS PROVIDES THIS CODE TO YOU WITHOUT WARRANTIES OF ANY KIND;
 *   (B) HORTONWORKS DISCLAIMS ANY AND ALL EXPRESS AND IMPLIED WARRANTIES WITH RESPECT TO THIS CODE, INCLUDING BUT NOT
 *     LIMITED TO IMPLIED WARRANTIES OF TITLE, NON-INFRINGEMENT, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE;
 *   (C) HORTONWORKS IS NOT LIABLE TO YOU, AND WILL NOT DEFEND, INDEMNIFY, OR HOLD YOU HARMLESS FOR ANY CLAIMS ARISING
 *     FROM OR RELATED TO THE CODE; AND
 *   (D) WITH RESPECT TO YOUR EXERCISE OF ANY RIGHTS GRANTED TO YOU FOR THE CODE, HORTONWORKS IS NOT LIABLE FOR ANY
 *     DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR CONSEQUENTIAL DAMAGES INCLUDING, BUT NOT LIMITED TO,
 *     DAMAGES RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF BUSINESS ADVANTAGE OR UNAVAILABILITY,
 *     OR LOSS OR CORRUPTION OF DATA.
 */

export class AssetCountModel {
  allCount: number;
  filesCount: number;
  hiveCount: number;
}

export class RichDatasetModel {
  counts: AssetCountModel;
  creatorId: number;
  creatorName: string;
  description: string;
  datalakeId: number; // datalakeId and dpClusterId are same thing
  datalakeName: string;
  favourite: boolean; // not related to favourite and bookmark (below)
  id: number;
  name: string;
  clusterId?: number;  // cluster id as discovered_cluster id
  tags?: string[];
  createdOn?: string;
  lastModified?: string;
  active?: boolean;
  version?: number;
  customProps?: any;
  sharedStatus?: number; // 1 public, 2 private
  favouriteId?: number;
  favouriteCount?: number;
  bookmarkId?: number;
  totalComments?: number;
  avgRating?: number;
  editDetails?: EditDetails;
}

export class Favourite {
  id: number;
  userId: number;
  objectType: string;
  objectId: number;
}

export class FavouriteWithTotal {
  favourite: Favourite;
  totalFavCount: number;
}

export class Bookmark {
  id: number;
  userId: number;
  objectType: string;
  objectId: number;
}

export class EditDetails {
  id: number;
  datasetId: number;
  editorId: number;
  editBegin: string
}